work on business cheks

+++ Shamane.Service/Implementation/Services/CenterService.cs
@@ -2,11 +2,15 @@
 using Shamane.DataAccess.UnitOfWorks;
 using Shamane.Domain;
 using Shamane.Domain.Conts;
+using Shamane.Domain.Exceptions;
+using Shamane.Service.Authentication.Service;
 using Shamane.Service.Definition;
 using Shamane.Service.Definition.Dto;
 using Shamane.Service.Definition.Factories;
+using Shamane.Service.Exceptions;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
 
 namespace Shamane.Service.Implementation.Services
@@ -15,23 +19,75 @@ namespace Shamane.Service.Implementation.Services
     {
         private readonly IUnitOfWork unitOfWork;
         private readonly ICenterFatory centerFatory;
-
-        public CenterService(IUnitOfWork unitOfWork, ICenterFatory centerFatory)
+        private readonly IAddressService addressService;
+        private readonly IUserService userService;
+        public CenterService(IUnitOfWork unitOfWork, 
+            ICenterFatory centerFatory,
+            IAddressService addressService,
+            IUserService userService)
         {
+            this.userService = userService;
+            this.addressService = addressService;
             this.unitOfWork = unitOfWork;
             this.centerFatory = centerFatory;
         }
+
+        private void centerValidator(CenterDto centerDto)
+        {
+            if (centerDto.CenterType == CenterType.Null ||
+                !centerDto.CenterType.IsDefined())
+            {
+                throw new CenterException(CenterExceptionRole.InvalidType);
+            }
+
+            if (centerDto.DeliveryType == DeliveryType.Null ||
+                !centerDto.DeliveryType.IsDefined())
+            {
+                throw new CenterException(CenterExceptionRole.InvalidDelivery);
+            }
+
+            if (string.IsNullOrEmpty(centerDto.Title) ||
+                string.IsNullOrWhiteSpace(centerDto.Title))
+            {
+                throw new CenterException(CenterExceptionRole.InvalidTitle);
+            }
+
+            if (string.IsNullOrEmpty(centerDto.CityId) ||
+                !centerDto.CityId.IsValidGuid() ||
+                !addressService.IsCityExistsAndActive(centerDto.CityId))
+            {
+                throw new AddressException(AddressExceptionRole.NotExistsOrDeactived);
+            }
+
+            //if (string.IsNullOrEmpty(centerDto.OwnerId) ||
+            //    !centerDto.OwnerId.IsValidGuid() ||
+            //    !addressService.IsCityExistsAndActive(centerDto.CityId))
+            //{
+            //    throw new AddressException(AddressExceptionRole.NotExistsOrDeactived);
+            //}
+
+        }
+
         public CenterDto Add(CenterDto centerDto)
         {
+            centerValidator(centerDto);
             var center = centerFatory.CreateEntity(centerDto);
             unitOfWork.CenterRepository.Add(center);
             unitOfWork.SaveChanges();
+            var province = addressService.GetProvince(center.City.ProvinceId.ToString());
+            center.City.Province = new Province()
+            {
+                Id = province.Id.ToGuid(),
+                Name = province.Name
+            };
             var addedDto = centerFatory.CreateDto(center);
             return addedDto;
         }
 
         public void Delete(string id)
         {
+            if (!id.IsValidGuid())
+            
             unitOfWork.CenterRepository.Remove(id.ToGuid());
             unitOfWork.SaveChanges();
         }
@@ -43,9 +99,11 @@ namespace Shamane.Service.Implementation.Services
         //    return centersDtos;
         //}
 
-        public CenterDto Get(Guid id)
+        public CenterDto Get(string id)
         {
-            var model = unitOfWork.CenterRepository.Get(id);
+            var model = unitOfWork.CenterRepository.Get(id.ToGuid());
+            if (model == null)
+                throw new EntityNotFoundException(typeof(CenterDto), id);
             var dto = centerFatory.CreateDto(model);
             return dto;
         }
@@ -60,12 +118,36 @@ namespace Shamane.Service.Implementation.Services
                 provinceId.ToNullableGuid(), cityId.ToNullableGuid(),
                 centerType, deliveryType, centerOrderBy,
                 from, count);
+
+            if (centers == null || centers.Count() == 0)
+                throw new EntityNotFoundException(typeof(CenterDto), "Query");
+
             var centersDtos = centerFatory.CreateDto(centers);
             return centersDtos;
         }
 
         public CenterDto Update(CenterDto centerDto)
         {
+            centerValidator(centerDto);
+
+            var center = centerFatory.CreateEntity(centerDto);
+            unitOfWork.CenterRepository.Update(center);
+            unitOfWork.SaveChanges();
+            var updatedDto = centerFatory.CreateDto(center);
+            return updatedDto;
+        }
+
+        public CenterDto Update(string id, CenterDto centerDto)
+        {
+            if (!id.IsValidGuid())
+            {
+                throw new EntityNotFoundException(typeof(CenterDto), id);
+            }
+            var centerForCheck = Get(id);
+            //if (centerForCheck.)
+            //{
+
+            //}
             throw new NotImplementedException();
         }
     }
